// interfaces/global.ts
export interface TerminalSession {
    STATUS: boolean;
    OUTPUT: string[];
    PROCESS?: any;
    PROCESS_PID?: string;
    LOCK?: boolean;
}

interface TerminalSessionsData {
    MESHCTL: TerminalSessionData;
    CONFIG: TerminalSessionData;
    CONTROLLER: TerminalSessionData;
    PROVISION: TerminalSessionData;
    SERVER: TerminalSessionData;
}

export interface TerminalSessions {
    MESHCTL: TerminalSession;
    CONFIG: TerminalSession;
    CONTROLLER: TerminalSession;
    PROVISION: TerminalSession;
    SERVER: TerminalSession;
}